chore: state language support and refactor
