+++ web/core/components/profile/overview/state-distribution.tsx
@@ -1,14 +1,12 @@
-// ui
+// plane imports
+import { STATE_GROUPS } from "@plane/constants";
 import { useTranslation } from "@plane/i18n";
 import { IUserProfileData, IUserStateDistribution } from "@plane/types";
 import { Card } from "@plane/ui";
+// components
 import { ProfileEmptyState, PieGraph } from "@/components/ui";
-
 // image
-import { STATE_GROUPS } from "@/constants/state";
 import stateGraph from "@/public/empty-state/state_graph.svg";
-// types
-// constants
 
 type Props = {
   stateDistribution: IUserStateDistribution[];
@@ -69,7 +67,7 @@ export const ProfileStateDistribution: React.FC<Props> = ({ stateDistribution, u
                           backgroundColor: STATE_GROUPS[group.state_group]?.color ?? "rgb(var(--color-primary-100))",
                         }}
                       />
-                      <div className="whitespace-nowrap">{STATE_GROUPS[group.state_group].label}</div>
+                      <div className="whitespace-nowrap">{t(STATE_GROUPS[group.state_group].i18n_label)}</div>
                     </div>
                     <div>{group.state_count}</div>
                   </div>
