+++ web/core/components/project-states/create-update/create.tsx
@@ -2,13 +2,14 @@
 
 import { FC, useState } from "react";
 import { observer } from "mobx-react";
+// plane imports
+import { STATE_GROUPS } from "@plane/constants";
 import { IState, TStateGroups } from "@plane/types";
 import { TOAST_TYPE, setToast } from "@plane/ui";
 // components
 import { StateForm } from "@/components/project-states";
 // constants
 import { STATE_CREATED } from "@/constants/event-tracker";
-import { STATE_GROUPS } from "@/constants/state";
 // hooks
 import { useEventTracker, useProjectState } from "@/hooks/store";
